pyproject.toml
[tool.poetry]
name = "collect-data"
version = "0.1.0"
description = ""
authors = ["munikumar"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
youtube-search-python = "^1.6.6"
yt-dlp = "^2023.11.16"
youtube-transcript-api = "^0.6.1"
pydub = "^0.25.1"
openai-whisper = "^20230918"
torch = {url = "https://download.pytorch.org/whl/cu118/torch-2.0.0%2Bcu118-cp39-cp39-linux_x86_64.whl" }
nemo-toolkit = {extras = ["asr"], version = "^1.19.0"}
torchaudio={url = "https://download.pytorch.org/whl/cu118/torchaudio-2.0.0%2Bcu118-cp39-cp39-linux_x86_64.whl"}
sqlmodel = "^0.0.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
